Profile is implemented in MVVM way, need fix photo problem. Regex added to validate email
