+++ SuppLocals/SuppLocals/Classes/ValidateUsername.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.ComponentModel;
 using System.Net.Mail;
+using System.Text.RegularExpressions;
 
 namespace SuppLocals
 {
@@ -90,15 +91,9 @@ namespace SuppLocals
 
         public bool IsValidEmail(string email)
         {
-            try
-            {
-                var addr = new MailAddress(email);
-                return addr.Address == email;
-            }
-            catch
-            {
-                return false;
-            }
+            
+                // Return true if strIn is in valid e-mail format.
+                return Regex.IsMatch(email, @"^([\w-\.]+)@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.)|(([\w-]+\.)+))([a-zA-Z]{2,4}|[0-9]{1,3})(\]?)$");
         }
     }
 }
