+++ SuppLocals/SuppLocals/Views/AccountViews/Login.xaml
@@ -5,8 +5,6 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
         xmlns:local="clr-namespace:SuppLocals"
         mc:Ignorable="d"
-        xmlns:viewModels="clr-namespace:SuppLocals.ViewModels"
-        xmlns:views="clr-namespace:SuppLocals.Views"
         Title="Login" Height="500" Width="400"
         AllowsTransparency="true" Background="Transparent"
         WindowStyle="none" ResizeMode="NoResize"
@@ -122,6 +120,8 @@
 
         </StackPanel>
 
+        
+        
     </Grid>
 
     <Window.Resources>
@@ -156,7 +156,6 @@
                 </Setter.Value>
             </Setter>
         </Style>
-
 
     </Window.Resources>
