--- SuppLocals/SuppLocals/Views/AccountViews/ProfileSettings.xaml
@@ -1,109 +1,0 @@
-﻿<Window x:Class="SuppLocals.Views.AccountViews.ProfileSettings"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:SuppLocals.Views.AccountViews"
-        mc:Ignorable="d"
-        Height="290" Width="500"
-        ResizeMode="NoResize"
-        WindowStyle="none" 
-        WindowStartupLocation="CenterScreen"
-        Background="#CCBA8B">
-    <Grid x:Name="grid">
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="auto"/>
-            <ColumnDefinition/>
-        </Grid.ColumnDefinitions>
-
-        <Grid.RowDefinitions>
-            <RowDefinition/>
-            <RowDefinition/>
-            <RowDefinition/>
-        </Grid.RowDefinitions>
-        
-        <StackPanel Grid.Row="0" Grid.Column="0" Grid.RowSpan="3">
-            <StackPanel Orientation="Horizontal">
-                
-                <Button Width="25" Height="25" Click="BackButtonClicked">
-                    <Image Source="/Assets/backarrow.png"/>
-                </Button>
-                
-                <TextBlock Text="Change your profile image" 
-                           HorizontalAlignment="Center" 
-                       Margin="10 9 0 0" FontSize="17"/>
-            </StackPanel>
-
-            <Button Margin="40" Background="Transparent" Click="ProfileImageClicked">
-                <Ellipse Width="150" Height="150">
-                    <Ellipse.Fill>
-                        <ImageBrush x:Name="profileImage"/>
-                    </Ellipse.Fill>
-                </Ellipse>
-
-
-            </Button>
-
-        </StackPanel>
-
-        <StackPanel Grid.Row="0" Grid.Column="1" Grid.RowSpan="3">
-            <TextBlock Text="Change your password" HorizontalAlignment="Center" 
-                       Margin="0 10 0 20 " FontSize="17"/>
-
-            <TextBlock Text="Old password" HorizontalAlignment="Left" 
-                       FontSize="13" Margin="10"/>
-
-            <Border CornerRadius="4" Width="200" BorderThickness="1" Background="White">
-                <PasswordBox x:Name="OldPass" Background="Transparent" BorderThickness="0"/>
-            </Border>
-
-            <TextBlock Text="New password" HorizontalAlignment="Left" 
-                       FontSize="13" Margin="10"/>
-
-            <Border CornerRadius="4" Width="200" BorderThickness="1" Background="White">
-                <PasswordBox x:Name="NewPass" Background="Transparent" BorderThickness="0"/>
-            </Border>
-
-            <TextBlock Text="Confirm new password" HorizontalAlignment="Left" 
-                       FontSize="13" Margin="10"/>
-
-            <Border CornerRadius="4" Width="200" BorderThickness="1" Background="White">
-                <PasswordBox x:Name="ConfirmNewPass" Background="Transparent" BorderThickness="0"/>
-            </Border>
-
-            <StackPanel Orientation="Horizontal" Margin="5">
-                <Grid Width="150">
-                    <Grid.RowDefinitions>
-                        <RowDefinition/>
-                        
-                    </Grid.RowDefinitions>
-
-                    <TextBlock x:Name="PasswordsDontMatch" Grid.Row="0" Text="Passwords don't match!" 
-                       Foreground="Red" FontSize="14"
-                       FontWeight="Bold" Visibility="Hidden"/>
-                    
-                    <TextBlock x:Name="EmptyFields" Grid.Row="0" 
-                       Text="Fields can't be empty!" Foreground="Red"
-                       FontSize="14" Margin="0 10 5 0" FontWeight="Bold"
-                       Visibility="Hidden"/>
-
-                    <TextBlock x:Name="OldMatchesNew" Grid.Row="0" Width="150" 
-                       Text="New password can't be your old!" Foreground="Red"
-                       FontSize="14" Margin="0 10 5 0" FontWeight="Bold"
-                       Visibility="Hidden" TextWrapping="Wrap"/>
-
-                    <TextBlock x:Name="IncorrectOldPass" Grid.Row="0" Width="150"
-                       Text="Old password is incorrect!" Foreground="Red"
-                       FontSize="14" Margin="0 10 5 0" FontWeight="Bold"
-                       Visibility="Hidden" TextWrapping="Wrap"/>
-                </Grid>
-
-                <Button HorizontalAlignment="Right" Margin="0 20"
-                    Content="Save changes" 
-                    Width="90" Height="25"
-                    Click="SaveChangesClicked"/>
-            </StackPanel>
-
-        </StackPanel>
-    </Grid>
-</Window>
